requirements.txt
absl-py==0.12.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bert-tensorflow==1.0.4
bertviz==1.0.0
bleach==3.3.0
boto3==1.17.45
botocore==1.20.45
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
entrypoints==0.3
fastprogress==1.0.0
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.34.1
h5py==3.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mxnet==1.8.0.post0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
NERDA==0.8.7
nest-asyncio==1.5.1
nltk==3.5
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
progressbar==2.5
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyconll==3.0.4
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.6
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2021032501
threadpoolctl==2.1.0
tokenizers==0.5.2
torch==1.7.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==2.5.1
transforms==0.1
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
